Server is feature complete.
All features specified in the brief now work, messages can be sent to
individuals or anyone, if you quit the clients information is removed
from server. Added 3 classes to test the client 2 are listeners that log
into the server and wait for output, the other sends commands to the
server to test all features work as expected.